+++ auto_gptq/quantization/__init__.py
@@ -1,2 +1,2 @@
 from .gptq import *
-from .quant import *
+from .quantizer import *
