+++ auto_gptq/quantization/gptq.py
@@ -7,7 +7,6 @@ import torch
 import torch.nn as nn
 import transformers
 
-from .quant import *
 
 logger = getLogger(__name__)
 
@@ -122,9 +121,7 @@ class GPTQ:
                         zero.append(self.quantizer.zero)
                         now_idx += 1
 
-                q = quantize(
-                    w.unsqueeze(1), self.quantizer.scale, self.quantizer.zero, self.quantizer.maxq
-                ).flatten()
+                q = self.quantizer.quantize(w.unsqueeze(1)).flatten()
                 Q1[:, i] = q
                 Losses1[:, i] = (w - q) ** 2 / d ** 2
 
@@ -176,5 +173,6 @@ class GPTQ:
         self.Losses = None
         self.Trace = None
         torch.cuda.empty_cache()
+
 
 __all__ = ["GPTQ"]
